Split monolith MCP server into 5 focused servers + add Affinity CRM integration
Break container/mcp-servers/google-calendar-gmail/ into 5 independent servers
(google, slack, lowercarbon, granola, affinity), each conditionally registered
based on credential existence. Add Affinity REST API v2 server for Deal Log
pipeline data, stage-aware NEW_DEAL context gathering, attachment surfacing
(Gmail download + Slack file download + send_file IPC), and per-attendee
Granola searches across all meeting types.
