+++ container/mcp-servers/affinity/tsconfig.json
