+++ docs/meeting-prep-agent.md
@@ -4,14 +4,15 @@ A NanoClaw agent that delivers contextual meeting briefings to Lowercarbon Capit
 
 ## Architecture
 
-The agent runs inside NanoClaw's Docker container with access to multiple data sources via MCP tools:
+The agent runs inside NanoClaw's Docker container with access to multiple data sources via 5 independent MCP servers:
 
-- **Google Calendar** — reads user's calendar + shared team calendars (OAuth, offline refresh token)
-- **Gmail** — searches email threads by address and/or name, returns thread links (OAuth, same token)
-- **Lowercarbon MCP** — portfolio documents, investor updates, IC memos, investment data (Vectorize, proxied via local stdio MCP server)
-- **Slack** — searches channels by company name, reads channel messages, auto-joins public channels (Bot token)
+- **Google** (`mcp-servers/google/`) — Calendar reads + Gmail thread/message search + attachment download (OAuth, offline refresh token)
+- **Slack** (`mcp-servers/slack/`) — Channel search, message search with file metadata, file download (Bot token)
+- **Lowercarbon** (`mcp-servers/lowercarbon/`) — Portfolio documents, investor updates, IC memos, investment data (Vectorize, JSON-RPC proxy)
+- **Granola** (`mcp-servers/granola/`) — Past meeting notes/transcripts search and retrieval (OAuth token)
+- **Affinity** (`mcp-servers/affinity/`) — Company search, Deal Log pipeline data, internal notes (REST API v2, Bearer token)
 
-All data access is read-only. The agent never sends emails, posts to channels, or writes to any system other than Slack DM delivery.
+Each server is conditionally registered — only starts if its credentials exist. All data access is read-only. The agent's only write capability is delivering briefings and file attachments to the user's Slack DM.
 
 ## Setup (per user)
 
@@ -27,7 +28,9 @@ Opens browser for OAuth consent. Token persists indefinitely (Internal Workspace
 Place in `groups/<group>/reference/`:
 - `google-credentials.json` — GCP OAuth Desktop client (from Cloud Console)
 - `google-token.json` — generated by auth script
-- `slack-bot-token.txt` — Slack bot token (`xoxb-...`)
+- `slack-bot-token.txt` — Slack bot token (`xoxb-...`, requires `files:read` and `files:write` scopes for attachment surfacing)
+- `affinity-api-key.txt` — Affinity API key (from Settings > Developer > API Keys)
+- `granola-token.json` — generated by Granola auth script (optional)
 
 ### MCP config
 Place in `groups/<group>/.mcp.json`:
@@ -57,17 +60,20 @@ The agent classifies calendar events before deciding what to research:
 | INTERNAL | All attendees on LC domains, no specific agenda | No briefing |
 | PORTFOLIO | Attendee domain or title matches portfolio company (confirmed via `get-portfolio-facts`) | Key updates + open threads |
 | LP | Attendee in LP CSV or title heuristics | Context + suggested discussion topics |
-| NEW_DEAL | Affinity Deal Log match or email inference (excludes investors/banks/corporates) | Full context + founder backgrounds |
+| NEW_DEAL | Affinity company search + Deal Log check, email inference (excludes investors/banks/corporates). Stage-aware context: early=heavy external, deep=heavy internal, pass=surface reason | Full context + founder backgrounds + deal pipeline data |
 | OTHER | Investors, banks, corporate partners | Lightweight brief |
 
-## Custom MCP Server
+## MCP Servers
 
-`container/mcp-servers/google-calendar-gmail/` is a local stdio MCP server built into the container image. It provides:
+Five independent MCP servers in `container/mcp-servers/`, each built into the container image:
 
-**Google Calendar tools:** `list_calendars`, `list_events`, `get_event`
-**Gmail tools:** `search_threads` (email + name union query), `get_thread`, `list_messages`
-**Slack tools:** `search_slack_channel`, `search_slack_messages` (with auto-join for public channels)
-**LC MCP proxy:** `portfolio_documents`, `get_company_facts`, `get_portfolio_facts` (proxies HTTP calls to Vectorize, bypassing OneCLI proxy)
+| Server | Tools | Auth |
+|--------|-------|------|
+| **google** | `list_calendars`, `list_events`, `get_event`, `search_threads`, `get_thread`, `list_messages`, `download_attachment` | Google OAuth (`google-credentials.json` + `google-token.json`) |
+| **slack** | `search_slack_channel`, `search_slack_messages`, `download_slack_file` | Bot token (`slack-bot-token.txt`) |
+| **lowercarbon** | `portfolio_documents`, `get_company_facts`, `get_portfolio_facts` | Vectorize API key (`.mcp.json`) |
+| **granola** | `query_granola_meetings`, `list_granola_meetings`, `get_granola_meeting` | OAuth token (`granola-token.json`) |
+| **affinity** | `search_affinity_companies`, `get_deal_log_entry`, `get_affinity_notes` | API key (`affinity-api-key.txt`) |
 
 ## GCP Project
 
@@ -83,18 +89,22 @@ The agent classifies calendar events before deciding what to research:
 | Salesforce MCP | LP prospects + older fund LPs (411, 419, 421, 7.81, n2O, SF6, CH4, Q>1) |
 | Hanover Park MCP | Existing LP fund admin (currently 430.5, Surge; historical migrating). URL: `app.hanoverpark.com/api/mcp` |
 | StandardMetrics | Financial metrics (needs company ID mapping from LC MCP names) |
-| Granola MCP | Past meeting notes/transcripts |
-| Affinity MCP | Deal pipeline, notes, relationship data |
+| iMessage / WhatsApp | Read-only messaging context (requires Google Contacts People API for phone number resolution) |
 
 ## Key Design Decisions
 
 - `get-portfolio-facts` (LC MCP) is the authority for company identification, not StandardMetrics
+- **NEW_DEAL meetings must NOT call Lowercarbon MCP** — it only contains portfolio company data
+- Affinity Deal Log (list ID 205572) is the authoritative source for deal pipeline status — field IDs are hardcoded
+- NEW_DEAL context gathering is stage-aware: early stage = heavy external, deep stage = heavy internal, pass = surface reason prominently
+- Granola searches per external attendee individually (all meeting types), not grouped
+- MCP servers are independent — 5 focused servers, each conditionally registered based on credential existence
 - Slack channels use shortened names vs LC MCP legal names — dynamic lookup via `search_slack_channel`, not a static mapping
 - Gmail search uses `{from:email from:"Name"}` union syntax to handle people with multiple email addresses
 - Calendar invite emails are filtered out of Gmail results
 - Briefings are concise (under 60 seconds to read), sources provided on demand
 - Agent asks for confirmation when uncertain about meeting classification
-- OneCLI proxy bypassed for remote MCP hosts via `NO_PROXY` env var
+- OneCLI proxy bypassed for remote MCP hosts via `NO_PROXY` env var (includes `api.affinity.co`)
 
 ## User Onboarding
